% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conflict_data.R
\docType{data}
\name{conflict_data}
\alias{conflict_data}
\title{Data from Armed Conflict Location & Event Data (ACLED) curated data from Latin America and the Caribbean.}
\format{
A tibble with 79417 obs. and 8 variables:
\describe{
\item{country}{country}
\item{location}{subnational location}
\item{event_type}{type of event}
\item{sub_event_type}{the specific event type}
\item{interaction}{what groups are interacting}
\item{fatalities}{the number of fatalaties recorded}
\item{event_date}{date of the event including day, month, and year}
\item{year}{the year of the event}

}

Raleigh, C., Kishi, R. & Linke, A. Political instability patterns are obscured by conflict dataset scope conditions, sources, and coding choices. Humanit Soc Sci Commun 10, 74 (2023). https://doi.org/10.1057/s41599-023-01559-4
}
\source{
\url{https://acleddata.com/latin-america-and-the-caribbean/#1649069068404-ec43820a-e533}
}
\usage{
conflict_data
}
\description{
The data is a simplified version of the ACLED Latin America & the Caribbean curated data, specifically it looks at Mexico and Colombia from 2018 to 2025 excluding demonstrations.
}
\details{
The data set was built with the intention to use of teaching POL 051: The Scientific Study of Politics.
}
\keyword{datasets}
