% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RandomData-package.R
\docType{package}
\name{RandomData-package}
\alias{RandomData}
\alias{RandomData-package}
\title{RandomData: POL 051 Data Package}
\description{
The following data sets are built for the use to help students learn to use R using a varity of different data that interests them. At the moment this includes: ANES 2008 data, ACLED Mexico and Colombia conflict data, F1 Data including on Drivers, Constructors, and Races.
}
\author{
\strong{Maintainer}: Haley Daarstad \email{hbdaarstad@ucdavis.edu}

}
\keyword{internal}
