% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RandomData-package.R
\docType{package}
\name{RandomData-package}
\alias{RandomData}
\alias{RandomData-package}
\title{RandomData: F1 Data}
\description{
F1 Data 1950 - 2024. F1 is a single-seater auto racing series sanctioned by the Fédération Internationale de l'Automobile (FIA) and owned by the Formula One Group.
}
\author{
\strong{Maintainer}: Haley Daarstad \email{hbdaarstad@ucdavis.edu}

}
\keyword{internal}
